requirements.txt
# ============================================================
# Customer Support Ticket Triage — OpenEnv Environment
# Dependencies for server-side Docker container
# ============================================================

openenv-core>=0.1.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
google-genai>=1.0.0
requests>=2.31.0
